+++ NEWS.md
@@ -1,3 +1,8 @@
+# conleyreg 0.1.6
+* `rowwise` option added
+* OpenMP optional
+* `par_dim` can be specified based on language
+
 # conleyreg 0.1.5
 * Major package extension
 * Package allows distances matrices to be pre-computed